Fixing gentoo init script to use system python.
If this is custom compiled on Gentoo (like it should be), the built in
python can have problems being found (especially if security mechanisms
are put in place (i.e. pax)).  This change tells nova-agent to use the
system python and check the system python's path before looking for the
nova-agent specific modules in the nova "shared" directory.