Make the log file no longer world readable
I've renamed a variable to make the code easier to understand.

Also it'll force the file permissions of the log file to be only
read|write by the owner, not by the whole world.

Just after the logging.basicConfig is called (which creates the log file
if it's not there already), it runs a chmod on the file.

This will happen even if the log file was made a long time ago.

-----

The other way to make the file read/writable but only when it's created
is to override the RotatingFileHandler's rotate method as shown here: http://stackoverflow.com/questions/1407474/does-python-logging-handlers-rotatingfilehandler-allow-creation-of-a-group-writa

However that's hard to do in C.

-----
Another way would be to check if the log file was created in the last
second, however that's not possible in posix C.